requirements.txt
adafruit-circuitpython-neopixel
rpi_ws281x

